[clang,tidy] fix a few warnings

+++ channels/cliprdr/cliprdr_common.c
@@ -388,7 +388,7 @@ UINT cliprdr_read_file_contents_response(wStream* s, CLIPRDR_FILE_CONTENTS_RESPO
 UINT cliprdr_read_format_list(wStream* s, CLIPRDR_FORMAT_LIST* formatList, BOOL useLongFormatNames)
 {
 	UINT32 index = 0;
-	int formatNameLength = 0;
+	size_t formatNameLength = 0;
 	const char* szFormatName = NULL;
 	const WCHAR* wszFormatName = NULL;
 	wStream sub1buffer = { 0 };
